fix: Fix stale index page by reporting the timestamp.
Instead of using zero as the timestamp, use the actual timestamp. This
should fix the issue of a stale page, when updating the UI. In that case
the index page is not fetched, and it tries to load the old resources
with the (no longer existing) cache busting markers. Resulting in a 404,
and resulting in a blank page.

Also, report error through result

Instead of rendering one error into the page and reporting the others
through a panic, report all through a result.