ci: create a "publish" branch, keeping published commits
Right now we have the issue that the tag points to a dangling commit,
as the commit itself doesn't belong to a branch. Once a new tag is
pushed, that commit might go away. This leads to failing local builds
and failing CI release builds.

The goal of this change is to create a "publish" branch per branch that
we publish, keeping the commits there.